move sdk connection to LoginFragment
